+++ source/soca/cluster_web_ui/views/my_files.py
@@ -327,7 +327,7 @@ def download():
                     total_files = total_files + 1
 
         if total_size > config.Config.MAX_ARCHIVE_SIZE:
-            flash("Sorry, the maximum archive size is {:.2f} MB. Your archive was {:.2f} MB. To mitigate this issue, you can create a smaller archive, download files individually, use SFTP or edit the maximum archive size authorized.".format(config.Config.MAX_ARCHIVE_SIZE/1024/1024, total_size/1024/1024), "error")
+            flash("Sorry, the maximum archive size is {:.2f} MB. Your archive was {:.2f} MB. To avoid this issue, you can create a smaller archive, download files individually, use SFTP or edit the maximum archive size authorized.".format(config.Config.MAX_ARCHIVE_SIZE/1024/1024, total_size/1024/1024), "error")
             return redirect("/my_files")
 
         # Limit HTTP payload size
@@ -365,9 +365,13 @@ def download_all():
         try:
             for entry in os.scandir(path):
                 if not entry.name.startswith("."):
-                    filesystem[entry.name] = {"path": path + "/" + entry.name,
+                    if entry.is_dir():
+                        # Ignore folder. We only include files
+                        pass
+                    else:
+                        filesystem[entry.name] = {"path": path + "/" + entry.name,
                                               "uid": encrypt(path + "/" + entry.name, entry.stat().st_size)["message"],
-                                              "type": "folder" if entry.is_dir() else "file",
+                                              "type": "file",
                                               "st_size": convert_size(entry.stat().st_size),
                                               "st_size_default": entry.stat().st_size,
                                               "st_mtime": entry.stat().st_mtime
@@ -400,7 +404,7 @@ def download_all():
         total_files = total_files + 1
 
     if total_size > config.Config.MAX_ARCHIVE_SIZE:
-        flash("Sorry, the maximum archive size is {:.2f} MB. Your archive was {:.2f} MB. To mitigate this issue, you can create a smaller archive, download files individually, use SFTP or edit the maximum archive size authorized.".format(config.Config.MAX_ARCHIVE_SIZE/1024/1024, total_size/1024/1024), "error")
+        flash("Sorry, the maximum archive size is {:.2f} MB. Your archive was {:.2f} MB. To avoid this issue, you can create a smaller archive, download files individually, use SFTP or edit the maximum archive size authorized.".format(config.Config.MAX_ARCHIVE_SIZE/1024/1024, total_size/1024/1024), "error")
         return redirect("/my_files")
 
     try: